==> Week 4/Part 02 - Segmenting and Clustering Neighbourhoods in Toronto - Applied Data Science Capstone - IBM Data Science Professional Certificate.ipynb <==
# Segmenting and Clustering Neighbourhoods in Toronto - Part 02
## Applied Data Science Capstone 
### IBM Data Science Professional Certificate

import pandas as pd
import geocoder
from operator import eq

### Load in the dataframe that was saved in the previous notebook

postalCodesDF = pd.read_csv("postalCodes.csv")
postalCodesDF.head()

*This took a prohibitively long time to run, so I'll use the CSV file provided instead*

### Use the Geospatial_Coordinates.csv to add the latitudes and longitudes to the postal codes dataframe

This Geocoding API took a prohibitively long time to run, so I'll use the CSV file provided instead. The geospatial coordinates file contains the postal codes in a different order to those in the postal codes dataframe from Part 01. Since we don't care about the ordering of the postal codes, the simplest way to incorporate them is simply to sort the 2 dataframes by the postal code column (ensuring that they are in the same order after the sort), then concatenate the latitude and longitude columns from the geospatial coordinates dataframe to the postal code, borough, and neighborhood columns from the postal codes dataframe.

geoCoordsDF  = pd.DataFrame(pd.read_csv("geospatial_Coordinates.csv"))

geoCoordsDF  .sort_values(by=["Postal Code"], inplace=True)
postalCodesDF.sort_values(by=["PostalCode"],  inplace=True)
postalCodesDF.reset_index(drop=True, inplace=True)

# This makes sure that when we concatenate the columns, the lat lon coords will match up with the postal code
assert(all(map(eq, geoCoordsDF["Postal Code"], postalCodesDF["PostalCode"])))

neighbourhoodsDF = pd.concat([postalCodesDF, geoCoordsDF[["Latitude", "Longitude"]]], axis=1)

neighbourhoodsDF.to_csv("neighbourhoods.csv", index=False)

neighbourhoodsDF.head()
